Migración de servidor Express a Vercel serverless. Borrado de recursos innecesarios
