+++ .gitignore
@@ -36,3 +36,4 @@ $RECYCLE.BIN/
 .DS_Store
 Thumbs.db
 UserInterfaceState.xcuserstate
+.vercel
