--- api/_certificates/cert.key
@@ -1,28 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQDIQOlZFznB8ZHZ
-EkWGCSBemQx60hPy2oI9wb4w/8fm1XrOdwgpam2qcKlRQg5Rz34KgueY/oMiYkEK
-qBmH6FuJiIRNffNU6ezLSgYMmh5wAgxkJc1h/uii3dCZ+kbvKwIzr2r4zXP1bF9g
-8BdP8g0CePlCIEjBwnuB9Uh++KXrkOx2P4N9RrN9aCPt3YPxdx3xRD7VhNlw8TKc
-a5b3ubrJ14BJA09mpBcHKITjW2S/EeFcivn7GCLFa4gkM9yZqgOLjIydd0CyNuHD
-x7ou7rliBX+B3wk4mAmI95ns9QFY2C5P3LtlaWW6bvvC/X/CayPmqGXUn6Avrb0u
-hMK4V39HAgMBAAECggEBAKi54KMbKaQa2Utv1LpZqDdsdxJpPTurFn/YZVq9uP3q
-s32Q14Sl0t7/6/m2UYBm9oj+wQqsr2riJHFJL7HOeooIR4lPbR4Ht2eV6EDn/mtb
-nnsnSsSd1f+cK64BLZFIyKMp01TtLx4DgN2ZjdrjVEjYCxTHROZU6mGWfM3WzV4L
-kF+18FCLGkz4y5Um8eNgIxFQowQvf0Y4pLsrFBWDoFATALGjldhF6ashLlsZMUko
-V3JcgXr8gjDPnhoWCvOOjgkYiDX8tZFip3zOiRy71V+gP8PF/ZmxW3kmo8EDZFG2
-MoLpy7zgSC3+eEXk4I85N+6fHgvdhJEkVz5zoccVwPECgYEA8wn3wIPmxBktD3Ef
-wkHXhvvTD6sHuaZRY8CNJK/lJmUD1M5pqD9plABCBLPNUcLRW2tqWhEkVCSSq058
-98yXmawFcIX7W0ypJqtmCoFARMG+F8meNkqGz6uCsQcMLgpsPumWXS7C3hMrn9XU
-WrIXKaxMeJMpzOd6RWlHkkbe29kCgYEA0u7TjtpfDhPSc0W0LaGTcOrmqm5YQrKp
-JAUd+AiWFNaGfnbUbB9j21D9iQHGlhU81bBHMWfmpPxKR71cQ0NeWsLZ2AlGVTM5
-J8NQXu9kwoYrlQrA5OHhFFtVzCRdu2b2cRpOemgzbfaqv0fZ44hBXOOlQK1w+7XU
-u3exGX3F4B8CgYEAqeL21xCcl8kNcXTfAVgrA0U2hQZFsSUsUXFAeDuoD42e/vuG
-acBh2b3QaB8TT6rB04ijX/AJl/O0vqBvxHu0IQxZBQ3y5SASQWCgRImS1j0j1tbt
-i//Oeo3mHj7R6cEJZ3S60colFz0JSoIbHYbNoCi7j7E3D+9izfBIwlxd3gkCgYAR
-XS23PczDaOhTKvLJKtrrWbsBXrIPU4vO9kjgSTypnHtX38pe2DqtLD6iz62cX0Ux
-xiAmNgprnIU04Ka+tsYGrLGoMiNDEkNrPY3ZMtXVmZvu2gLFFL8loNswQfIhYjl5
-WBQLyZYqQDsQIgcSzv2qeub4qw54+h/4CsNeeNcYOwKBgE2AGt2piOenJYyJ+0mJ
-1/xDjQL91iHEESSnKHZq9DPi1dawMEoo6+A2LgJQyXV4DY/jbrywKeL5WuDh83ps
-wgtRlv7vP/F88TN7vEZOdC88Vo9K9PYf5cR5kAtfsPPv/m6wzfNuzWmCF4qe1ALS
-+6yeSPuQWxsjHcuFwiNAW9mV
------END PRIVATE KEY-----
