--- api/_certificates/cert.pem
@@ -1,24 +1,0 @@
------BEGIN CERTIFICATE-----
-MIID+zCCAuOgAwIBAgIUeV9NOaJktOzLKSVDR8xNZlJudTwwDQYJKoZIhvcNAQEL
-BQAwgYYxCzAJBgNVBAYTAkFSMQ4wDAYDVQQIDAVTdGF0ZTERMA8GA1UEBwwITG9j
-YXRpb24xGjAYBgNVBAoMEU9yZ2FuaXphdGlvbiBOYW1lMRwwGgYDVQQLDBNPcmdh
-bml6YXRpb25hbCBVbml0MRowGAYDVQQDDBF3d3cubG9jYWxob3N0LmNvbTAeFw0y
-MjAxMjQwNTIyNTZaFw0yMzAxMjQwNTIyNTZaMIGGMQswCQYDVQQGEwJBUjEOMAwG
-A1UECAwFU3RhdGUxETAPBgNVBAcMCExvY2F0aW9uMRowGAYDVQQKDBFPcmdhbml6
-YXRpb24gTmFtZTEcMBoGA1UECwwTT3JnYW5pemF0aW9uYWwgVW5pdDEaMBgGA1UE
-AwwRd3d3LmxvY2FsaG9zdC5jb20wggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEK
-AoIBAQDIQOlZFznB8ZHZEkWGCSBemQx60hPy2oI9wb4w/8fm1XrOdwgpam2qcKlR
-Qg5Rz34KgueY/oMiYkEKqBmH6FuJiIRNffNU6ezLSgYMmh5wAgxkJc1h/uii3dCZ
-+kbvKwIzr2r4zXP1bF9g8BdP8g0CePlCIEjBwnuB9Uh++KXrkOx2P4N9RrN9aCPt
-3YPxdx3xRD7VhNlw8TKca5b3ubrJ14BJA09mpBcHKITjW2S/EeFcivn7GCLFa4gk
-M9yZqgOLjIydd0CyNuHDx7ou7rliBX+B3wk4mAmI95ns9QFY2C5P3LtlaWW6bvvC
-/X/CayPmqGXUn6Avrb0uhMK4V39HAgMBAAGjXzBdMA4GA1UdDwEB/wQEAwIDiDAT
-BgNVHSUEDDAKBggrBgEFBQcDATA2BgNVHREELzAtghF3d3cubG9jYWxob3N0LmNv
-bYINbG9jYWxob3N0LmNvbYIJbG9jYWxob3N0MA0GCSqGSIb3DQEBCwUAA4IBAQCk
-kwJ5OpECxwTXbNqauf+u0Kneya6uOY9AZ1qGo11UxCtOMWpIrBsxcTThpIqhQFJ3
-vFqM+P0HI93UafJU1H7wzmZg2zFVu3u/PchdUyYdYj8B/RGE8L/V+OUFYSfOaGe4
-lfwVxBY+4R9n6eQc7EuRK5+/ImJZKt5ooyyb1ONcv3dvhcwQljRRcdFsRjBywFyw
-638jCWr4bSPpEFQVjdJTnAkELd8u4xXF6NXqNF3rTWGH29wjX2FDKdTpuNcIr8+5
-dtFsLEHYZfsloQvBnEUUJObUPOZCeJhP6jrMxChUPc3wBbw1BsVqMKNr/ittytEP
-qvOADieakvdasgHiknhs
------END CERTIFICATE-----
