--- api/controllers/push-notifications.controller.js
@@ -1,14 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-const pushNotificationsService = require('../services/push-notifications.service');
-router.get('/getAppId', getAppId);
-
-module.exports = router;
-
-function getAppId(req, res, next) {
-    return pushNotificationsService
-        .getAppId()
-        .then((result) => res.json(result))
-        .catch((err) => next(err));
-}
