+++ api/push-notifications/get-app-id.js
@@ -1,0 +1,5 @@
+const environment = require('../_helpers/environment');
+
+export default function getAppId(req, res) {
+    res.json(environment.oneSignalAppId);
+}
