--- api/services/push-notifications.service.js
@@ -1,11 +1,0 @@
-const environment = require('../_helpers/environment');
-
-module.exports = {
-    getAppId,
-};
-
-async function getAppId() {
-    return new Promise((resolve, reject) => {
-        resolve(environment.oneSignalAppId);
-    });
-}
