--- api/services/story.service.js
@@ -1,70 +1,0 @@
-const sanityConnector = require('../_helpers/sanity-connector');
-const imageUrlBuilder = require('@sanity/image-url');
-
-const builder = imageUrlBuilder(sanityConnector.client);
-
-function urlFor(source) {
-    return builder.image(source);
-}
-
-module.exports = {
-    getAuthors,
-    getById,
-    getCount,
-    getOriginalLinks,
-};
-
-async function getAuthors() {
-    const query = `*[_type == 'story'] | order(day desc) {title, day, categories, publishedAt, author->}`;
-    const authors = await sanityConnector.client.fetch(query, {});
-    return authors.map((x) => ({ ...x, author: mapAuthor(x.author) }));
-}
-
-async function getById(id) {
-    const query = `*[_type == 'story' && day == ${id}]{title, day, originalLink, forewords, categories, publishedAt, body, review, forewords, author->}`;
-    const result = await sanityConnector.client.fetch(query, {});
-
-    let story = result.length ? result.pop() : null;
-
-    if (story) {
-        story = {
-            ...story,
-            summary: story.review,
-            paragraphs: mapBodyToParagraphs(story.body),
-            author: mapAuthor(story.author),
-            prologues: mapPrologues(story.forewords),
-        };
-    }
-
-    return story;
-}
-
-async function getCount() {
-    const query = `count(*[_type == 'story'])`;
-    const result = await sanityConnector.client.fetch(query, {});
-    return result;
-}
-
-async function getOriginalLinks() {
-    const query = `*[_type == 'story'] | order(day desc) {title, day, author->, originalLink}`;
-    const result = await sanityConnector.client.fetch(query, {});
-    return result;
-}
-
-function mapAuthor(authorDTO) {
-    return {
-        biography: authorDTO.bio,
-        nationality: urlFor(authorDTO.country).url(),
-        fullBioUrl: authorDTO.fullBioUrl,
-        imageUrl: urlFor(authorDTO.image).url(),
-        name: authorDTO.name,
-    };
-}
-
-function mapBodyToParagraphs(story) {
-    return story;
-}
-
-function mapPrologues(prologuesDTO) {
-    return prologuesDTO ? prologuesDTO.map((x) => ({ reference: x.fwAuthor, text: x.fwText })) : [];
-}
