+++ src/environments/environment.ts
@@ -4,7 +4,7 @@
 // The list of which env maps to which file can be found in `.angular-cli.json`.
 export const environment = {
     production: false,
-    apiUrl: 'http://localhost:8100',
+    apiUrl: 'http://localhost:3000',
 };
 
 /*
